Factor out batching stuff from server/service/kv
