+++ src/server/service/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2017 TiKV Project Authors. Licensed under Apache-2.0.
 
+mod batch;
 mod debug;
 mod diagnostics;
 mod kv;
